# docs/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Layer para integrações de pagamento Afya.",
        "title": "Card Layer API",
        "contact": {
            "name": "Foundation Squad"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/cards": {
            "post": {
                "security": [
                    {
                        "x-api-key": []
                    }
                ],
                "description": "Cria um cartão e o tokemiza nos gateways de pagamento",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cards"
                ],
                "summary": "Criar um cartão",
                "parameters": [
                    {
                        "description": "Informações sobre o cartão e informações do pagador",
                        "name": "card",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/payment-layer-card-api_usecases_cards_dtos.CreateCardDTOInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/payment-layer-card-api_usecases_cards_dtos.CreateCardDTOOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/adapters_http.ErrorWrapperInternalDTO"
                        }
                    }
                }
            }
        },
        "/cards/{cardId}/customer/{customerId}": {
            "get": {
                "security": [
                    {
                        "x-api-key": []
                    }
                ],
                "description": "Buscar um cartão pelo identificador e id do cliente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cards"
                ],
                "summary": "Obter um cartão pelo id e id do cliente",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Identificador do cartão",
                        "name": "cardId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Identificador do cliente",
                        "name": "customerId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/payment-layer-card-api_usecases_cards_dtos.GetCardWithoutTokensDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/adapters_http.ErrorWrapperInternalDTO"
                        }
                    }
                }
            }
        },
        "/cards/{id}": {
            "get": {
                "security": [
                    {
                        "x-api-key": []
                    }
                ],
                "description": "Buscar um cartão pelo identificador",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cards"
                ],
                "summary": "Obter um cartão pelo id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Identificador do cartão",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/payment-layer-card-api_usecases_cards_dtos.GetCardWithTokensDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/adapters_http.ErrorWrapperInternalDTO"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "adapters_http.ErrorOutput": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "http_status": {
                    "type": "integer"
                },
                "log_message": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "adapters_http.ErrorWrapperInternalDTO": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/adapters_http.ErrorOutput"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "payment-layer-card-api_usecases_cards_dtos.BillingAddress": {
            "type": "object",
            "required": [
                "city",
                "country",
                "line_1",
                "state"
            ],
            "properties": {
                "city": {
                    "description": "Cidade do endereço de cobrança",
                    "type": "string",
                    "example": "São Paulo"
                },
                "country": {
                    "description": "País do endereço de cobrança",
                    "type": "string",
                    "example": "BR"
                },
                "district": {
                    "type": "string",
                    "example": "Centro"
                },
                "line_1": {
                    "description": "Linha 1 do endereço de cobrança",
                    "type": "string",
                    "example": "Rua João Silva, 123"
                },
                "line_2": {
                    "description": "Linha 2 do endereço de cobrança",
                    "type": "string",
                    "example": "Apto 123"
                },
                "postalCode": {
                    "description": "CEP do endereço de cobrança",
                    "type": "string",
                    "example": "12345678"
                },
                "postal_code": {
                    "description": "TODO: MUDAR PARA SNAKE CASE                              // Bairro do endereço de cobrança",
                    "type": "string",
                    "example": "12345678"
                },
                "state": {
                    "description": "Estado do endereço de cobrança",
                    "type": "string",
                    "example": "SP"
                }
            }
        },
        "payment-layer-card-api_usecases_cards_dtos.CardToken": {
            "type": "object",
            "properties": {
                "gateway": {
                    "description": "Gateway do token",
                    "type": "string",
                    "example": "rede"
                },
                "token": {
                    "description": "Token do cartão",
                    "type": "string",
                    "example": "9KwUaB1kwz8="
                }
            }
        },
        "payment-layer-card-api_usecases_cards_dtos.CreateCardDTOInput": {
            "type": "object",
            "required": [
                "billing_address",
                "cvv",
                "exp_month",
                "exp_year",
                "holder",
                "number"
            ],
            "properties": {
                "billing_address": {
                    "description": "Endereço de cobrança do cartão",
                    "allOf": [
                        {
                            "$ref": "#/definitions/payment-layer-card-api_usecases_cards_dtos.BillingAddress"
                        }
                    ]
                },
                "brand": {
                    "description": "Bandeira do cartão",
                    "type": "string",
                    "example": "visa"
                },
                "customer": {
                    "description": "Cliente do cartão",
                    "allOf": [
                        {
                            "$ref": "#/definitions/payment-layer-card-api_usecases_cards_dtos.Customer"
                        }
                    ]
                },
                "customer_id": {
                    "description": "Identificador do cliente",
                    "type": "string",
                    "example": "835524dd-d897-45ed-a68e-2e1a8bf52314"
                },
                "cvv": {
                    "description": "Código de segurança do cartão",
                    "type": "string",
                    "example": "123"
                },
                "exp_month": {
                    "description": "Mês de expiração do cartão",
                    "type": "string",
                    "example": "12"
                },
                "exp_year": {
                    "description": "Ano de expiração do cartão",
                    "type": "string",
                    "example": "25"
                },
                "holder": {
                    "description": "Nome do portador do cartão",
                    "type": "string",
                    "example": "Joao Silva"
                },
                "number": {
                    "description": "Número do cartão",
                    "type": "string",
                    "example": "4111111111111111"
                }
            }
        },
        "payment-layer-card-api_usecases_cards_dtos.CreateCardDTOOutput": {
            "type": "object",
            "properties": {
                "customer_id": {
                    "description": "Identificador do cliente",
                    "type": "string",
                    "example": "835524dd-d897-45ed-a68e-2e1a8bf52314"
                },
                "id": {
                    "description": "Identificador do cartão",
                    "type": "string",
                    "example": "835524dd-d897-45ed-a68e-2e1a8bf522d7"
                }
            }
        },
        "payment-layer-card-api_usecases_cards_dtos.Customer": {
            "type": "object",
            "required": [
                "address",
                "birth_date",
                "document",
                "email",
                "first_name",
                "gender",
                "last_name",
                "phone",
                "service_id"
            ],
            "properties": {
                "address": {
                    "$ref": "#/definitions/payment-layer-card-api_usecases_cards_dtos.BillingAddress"
                },
                "birth_date": {
                    "type": "string",
                    "example": "2000-11-06"
                },
                "document": {
                    "type": "string",
                    "example": "12345678901"
                },
                "email": {
                    "type": "string",
                    "example": "teste@gmail.com"
                },
                "first_name": {
                    "type": "string",
                    "example": "João"
                },
                "gender": {
                    "type": "string",
                    "enum": [
                        "male",
                        "female"
                    ],
                    "example": "male || female"
                },
                "last_name": {
                    "type": "string",
                    "example": "Silva"
                },
                "phone": {
                    "type": "string",
                    "example": "11999999999"
                },
                "service_id": {
                    "type": "string",
                    "example": "835524dd-d897-45ed-a68e-2e1a8bf52314"
                }
            }
        },
        "payment-layer-card-api_usecases_cards_dtos.GetCardWithTokensDTO": {
            "type": "object",
            "properties": {
                "brand": {
                    "description": "Bandeira do cartão",
                    "type": "string",
                    "example": "visa"
                },
                "created_at": {
                    "description": "Data de criação do cartão",
                    "type": "string",
                    "example": "2006-01-02T15:04:05-07:00"
                },
                "customer_id": {
                    "description": "Identificador do cliente",
                    "type": "string",
                    "example": "835524dd-d897-45ed-a68e-2e1a8bf52314"
                },
                "exp_month": {
                    "description": "Mês de expiração do cartão",
                    "type": "string",
                    "example": "12"
                },
                "exp_year": {
                    "description": "Ano de expiração do cartão",
                    "type": "string",
                    "example": "2025"
                },
                "first_digits": {
                    "description": "Primeiros dígitos do cartão",
                    "type": "string",
                    "example": "411111"
                },
                "holder": {
                    "description": "Nome do portador do cartão",
                    "type": "string",
                    "example": "Joao Silva"
                },
                "id": {
                    "description": "Identificador do cartão",
                    "type": "string",
                    "example": "835524dd-d897-45ed-a68e-2e1a8bf522d7"
                },
                "last_digits": {
                    "description": "Últimos dígitos do cartão",
                    "type": "string",
                    "example": "1111"
                },
                "tokens": {
                    "description": "Tokens do cartão",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/payment-layer-card-api_usecases_cards_dtos.CardToken"
                    }
                },
                "updated_at": {
                    "description": "Data de atualização do cartão",
                    "type": "string",
                    "example": "2006-01-02T15:04:05-07:00"
                }
            }
        },
        "payment-layer-card-api_usecases_cards_dtos.GetCardWithoutTokensDTO": {
            "type": "object",
            "properties": {
                "brand": {
                    "description": "Bandeira do cartão",
                    "type": "string",
                    "example": "visa"
                },
                "created_at": {
                    "description": "Data de criação do cartão",
                    "type": "string",
                    "example": "2006-01-02T15:04:05-07:00"
                },
                "customer_id": {
                    "description": "Identificador do cliente",
                    "type": "string",
                    "example": "835524dd-d897-45ed-a68e-2e1a8bf52314"
                },
                "exp_month": {
                    "description": "Mês de expiração do cartão",
                    "type": "string",
                    "example": "12"
                },
                "exp_year": {
                    "description": "Ano de expiração do cartão",
                    "type": "string",
                    "example": "2025"
                },
                "first_digits": {
                    "description": "Primeiros dígitos do cartão",
                    "type": "string",
                    "example": "411111"
                },
                "holder": {
                    "description": "Nome do portador do cartão",
                    "type": "string",
                    "example": "Joao Silva"
                },
                "id": {
                    "description": "Identificador do cartão",
                    "type": "string",
                    "example": "835524dd-d897-45ed-a68e-2e1a8bf522d7"
                },
                "last_digits": {
                    "description": "Últimos dígitos do cartão",
                    "type": "string",
                    "example": "1111"
                },
                "updated_at": {
                    "description": "Data de atualização do cartão",
                    "type": "string",
                    "example": "2006-01-02T15:04:05-07:00"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "x-api-key": {
            "type": "apiKey",
            "name": "x-api-key",
            "in": "header"
        }
    }
}